+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Fix bug that Slimi::Interpolation was not working.
+
 ## 0.1.0 - 2021-12-20
 
 - Initial release.